Core & Internals: move checksum globals & friends out of common.utils #5079
CHECKSUM_KEY is now defined in core.rse, as it is the only module
which actually uses it. The rest has been relocated to
common.hash_algorithms.
