+++ lib/rucio/rse/protocols/ssh.py
@@ -21,7 +21,8 @@ import os
 import re
 
 from rucio.common import exception
-from rucio.common.utils import execute, PREFERRED_CHECKSUM
+from rucio.common.hash_algorithms import PREFERRED_CHECKSUM
+from rucio.common.utils import execute
 from rucio.rse.protocols import protocol
