+++ lib/rucio/rse/protocols/xrootd.py
@@ -30,8 +30,9 @@ import os
 import logging
 
 from rucio.common import exception
+from rucio.common.hash_algorithms import PREFERRED_CHECKSUM
+from rucio.common.utils import execute
 from rucio.rse.protocols import protocol
-from rucio.common.utils import execute, PREFERRED_CHECKSUM
 
 
 class Default(protocol.RSEProtocol):
