+++ lib/rucio/rse/rsemanager.py
@@ -56,8 +56,9 @@ except ImportError:
 from rucio.common import exception, utils, constants
 from rucio.common.config import config_get_int
 from rucio.common.constraints import STRING_TYPES
+from rucio.common.hash_algorithms import GLOBALLY_SUPPORTED_CHECKSUMS
 from rucio.common.logging import formatted_logger
-from rucio.common.utils import make_valid_did, GLOBALLY_SUPPORTED_CHECKSUMS
+from rucio.common.utils import make_valid_did
 
 
 def get_rse_info(rse=None, vo='def', rse_id=None, session=None):
